Event display updates
Updated the display of events to show the doctor’s name and their job
title on the event itself. Updated the shift manager to save the data
from the scheduleNotes text area to the event as well. Todo: Save the
date textbox to the event, and change the names of variables to be
appropriate for their content.